requirements.txt
streamlit
ultralytics
opencv-python
python-dotenv
google-generativeai